Change Directory structure to simplify
