--- src/components/Continents/Continents.js
@@ -1,38 +1,0 @@
-import React, { useEffect } from 'react';
-import { Link } from 'react-router-dom';
-import { useDispatch, useSelector } from 'react-redux';
-import { getAllData } from '../../redux/continent';
-import { GridContainer, Button } from './ContinentsStyle';
-
-function Continents() {
-  const dispatch = useDispatch();
-  const continents = useSelector((state) => state.reduContinent);
-  useEffect(() => {
-    dispatch(getAllData);
-  }, []);
-  return (
-    <GridContainer>
-      {continents.map((continent) => (
-        <ul key={continent.id} style={{ listStyle: 'none' }}>
-          <li><h2>{continent.continent}</h2></li>
-          <li>
-            <strong>Cases:</strong>
-            {continent.cases}
-          </li>
-          <li>
-            <strong>Deaths:</strong>
-            {continent.deaths}
-          </li>
-          <li>
-            {/* {console.log(continent.countries)} */}
-            <Link to={`/country/${continent.id}`}>
-              <Button>Get Info</Button>
-            </Link>
-          </li>
-        </ul>
-      ))}
-    </GridContainer>
-  );
-}
-
-export default Continents;
