+++ src/components/Styles/ContinentsStyle.js
