--- src/components/pages/Home.js
@@ -1,13 +1,0 @@
-import React from 'react';
-import Continents from '../Continents/Continents';
-
-function Home() {
-  return (
-    <div>
-      <h1>Covid-19 Metrics</h1>
-      <Continents />
-    </div>
-  );
-}
-
-export default Home;
