+++ src/pages/Country.js
@@ -1,25 +1,24 @@
-import PropTypes from 'prop-types';
+import { useParams } from 'react-router-dom';
 import React, { useEffect, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { getAllData } from '../../redux/continent';
-import { Button } from './CountriesStyle';
-import CountryInfo from '../CountryInfo/CountryInfo';
+import { getAllData } from '../redux/continent';
+import { Button } from '../components/Styles/CountriesStyle';
+import CountryInfo from '../components/CountryInfo/CountryInfo';
 
-function Countries(props) {
+function Country() {
   const dispatch = useDispatch();
   useEffect(() => {
     dispatch(getAllData);
   }, []);
-
+  const { continent } = useParams();
   const continents = useSelector((state) => state.reduContinent);
-  const { name } = props;
-  const filtercontinent = continents.filter((continent) => (continent.id === name))[0];
+  const filtercontinent = continents.filter((item) => (item.id === continent))[0];
   const { countries } = filtercontinent;
   const [country, setCountry] = useState(countries[0]);
 
   return (
     <div>
-      <h1>{name}</h1>
+      <h1>{continent}</h1>
       <form onSubmit={() => setCountry()}>
         <select name="country">
           {countries.map((country) => <option value={`${country}`} key={country.split(3)}>{country}</option>)}
@@ -31,8 +30,4 @@ function Countries(props) {
   );
 }
 
-Countries.propTypes = {
-  name: PropTypes.string.isRequired,
-};
-
-export default Countries;
+export default Country;
