+++ src/pages/Home.js
@@ -1,0 +1,41 @@
+import React, { useEffect } from 'react';
+import { Link } from 'react-router-dom';
+import { useDispatch, useSelector } from 'react-redux';
+import { getAllData } from '../redux/continent';
+import { GridContainer, Button } from '../components/Styles/ContinentsStyle';
+
+function Home() {
+  const dispatch = useDispatch();
+  const continents = useSelector((state) => state.reduContinent);
+  useEffect(() => {
+    dispatch(getAllData);
+  }, []);
+  return (
+    <div>
+      <h1>Covid-19 Metrics</h1>
+      <GridContainer>
+        {continents.map((continent) => (
+          <ul key={continent.id} style={{ listStyle: 'none' }}>
+            <li><h2>{continent.continent}</h2></li>
+            <li>
+              <strong>Cases:</strong>
+              {continent.cases}
+            </li>
+            <li>
+              <strong>Deaths:</strong>
+              {continent.deaths}
+            </li>
+            <li>
+              {/* {console.log(continent.countries)} */}
+              <Link to={`/country/${continent.id}`}>
+                <Button>Get Info</Button>
+              </Link>
+            </li>
+          </ul>
+        ))}
+      </GridContainer>
+    </div>
+  );
+}
+
+export default Home;
